# services/book-tracking-service/src/generated/openapi/com/tranvodev/book_tracking_service/api/BooksApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.22.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.tranvodev.book_tracking_service.api;

import com.tranvodev.book_tracking_service.dto.ErrorResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.22.0")
@Validated
@Tag(name = "Books", description = "the Books API")
public interface BooksApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_UPLOAD_BOOK = "/books";
    /**
     * POST /books : Upload a book
     * Uploads a book file and creates a new book record. The service forwards the file to the external file storage service&#39;s /files endpoint and stores the returned file metadata against the new book.
     *
     * @param file The file to upload. (required)
     * @return Book uploaded and created. (status code 201)
     *         or Invalid request (e.g. missing file or unsupported type). (status code 400)
     *         or File too large. (status code 413)
     *         or Internal server error. (status code 500)
     */
    @Operation(
        operationId = "uploadBook",
        summary = "Upload a book",
        description = "Uploads a book file and creates a new book record. The service forwards the file to the external file storage service's /files endpoint and stores the returned file metadata against the new book.",
        tags = { "Books" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Book uploaded and created."),
            @ApiResponse(responseCode = "400", description = "Invalid request (e.g. missing file or unsupported type).", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "413", description = "File too large.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = BooksApi.PATH_UPLOAD_BOOK,
        produces = { "application/json" },
        consumes = { "multipart/form-data" }
    )
    default ResponseEntity<Void> uploadBook(
        @Parameter(name = "file", description = "The file to upload.", required = true) @RequestPart(value = "file", required = true) MultipartFile file
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"instance\" : \"https://openapi-generator.tech\", \"detail\" : \"detail\", \"type\" : \"https://openapi-generator.tech\", \"title\" : \"title\", \"status\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"instance\" : \"https://openapi-generator.tech\", \"detail\" : \"detail\", \"type\" : \"https://openapi-generator.tech\", \"title\" : \"title\", \"status\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"instance\" : \"https://openapi-generator.tech\", \"detail\" : \"detail\", \"type\" : \"https://openapi-generator.tech\", \"title\" : \"title\", \"status\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}