# services/book-tracking-service/src/main/resources/openapi/book-app-api.yaml
openapi: 3.0.3
info:
  title: Book App API
  version: 1.0.0
  description: API for the Book App, including file upload support.

servers:
  - url: http://localhost:8080/book-tracking/api

paths:
  /books:
    post:
      summary: Upload a book
      description: >-
        Uploads a book file and creates a new book record. The service
        forwards the file to the external file storage service's /files
        endpoint and stores the returned file metadata against the new book.
      operationId: uploadBook
      tags:
        - Books
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileUploadRequest'
            encoding:
              file:
                contentType: application/octet-stream
      responses:
        '201':
          description: Book uploaded and created.
        '400':
          description: Invalid request (e.g. missing file or unsupported type).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: File too large.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    FileUploadRequest:
      type: object
      required:
        - file
      properties:
        file:
          type: string
          format: binary
          description: The file to upload.

    ErrorResponse:
      type: object
      description: A Problem Details object as defined by RFC 9457.
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type.
          default: about:blank
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          description: The HTTP status code generated by the origin server for this occurrence of the problem.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence of the problem.
      required:
        - type
        - title
        - status